optimize esp32 send

+++ components/micropython/port/src/standard_lib/include/esp32_spi.h
@@ -167,7 +167,7 @@ uint8_t esp32_spi_get_socket(void);
 int8_t esp32_spi_socket_open(uint8_t sock_num, uint8_t *dest, uint8_t dest_type, uint16_t port, esp32_socket_mode_enum_t conn_mode);
 esp32_socket_enum_t esp32_spi_socket_status(uint8_t socket_num);
 uint8_t esp32_spi_socket_connected(uint8_t socket_num);
-uint32_t esp32_spi_socket_write(uint8_t socket_num, uint8_t *buffer, uint32_t len);
+uint32_t esp32_spi_socket_write(uint8_t socket_num, uint8_t *buffer, uint16_t len);
 int esp32_spi_socket_available(uint8_t socket_num);
 int esp32_spi_socket_read(uint8_t socket_num, uint8_t *buff, uint16_t size);
 int8_t esp32_spi_socket_connect(uint8_t socket_num, uint8_t *dest, uint8_t dest_type, uint16_t port, esp32_socket_mode_enum_t conn_mod);
